+++ Controllers/PostController.js
@@ -19,7 +19,15 @@ export const createPost = async (req, res) => {
 };
 
 export const getPosts = async (req, res) => {
-	const posts = await prisma.post.findMany();
+	const posts = await prisma.post.findMany({
+		include: {
+			_count: {
+				select: {
+					Comments: true,
+				},
+			},
+		},
+	});
 
 	return res.json({
 		status: 200,
@@ -28,12 +36,32 @@ export const getPosts = async (req, res) => {
 	});
 };
 
+export const getPostsByUser = async (req, res) => {
+	const { user_id } = req.params;
+
+	const posts = await prisma.post.findMany({
+		where: {
+			user_id: parseInt(user_id),
+		},
+	});
+
+	return res.json({
+		status: 200,
+		data: posts,
+		message: "All posts by the user are fetched.",
+	});
+};
+
+// Get a post by its ID.
 export const getPostById = async (req, res) => {
 	const { id } = req.params;
 
 	const post = await prisma.post.findUnique({
 		where: {
 			id: parseInt(id),
+		},
+		include: {
+			Comments: true,
 		},
 	});
 
@@ -67,6 +95,19 @@ export const updatePost = async (req, res) => {
 
 export const deletePost = async (req, res) => {
 	const { id } = req.params;
+
+	const post = await prisma.post.findUnique({
+		where: {
+			id: parseInt(id),
+		},
+	});
+
+	if (!post) {
+		return res.json({
+			status: 404,
+			message: "Post not found.",
+		});
+	}
 
 	const deletedPost = await prisma.post.delete({
 		where: {
